+++ src/Modules/Admin/Widgets/Forms/AccountResendConfirmActiveForm.php
@@ -35,14 +35,18 @@ class AccountResendConfirmActiveForm extends ActiveForm
     {
         $this->attributes['id'] ??= 'resend-form';
 
-        $this->rows ??= [
-            $this->getHelpText(),
-            $this->getEmailField(),
-        ];
-
         $this->submitButtonText ??= Yii::t('skeleton', 'COMMON_SEND_EMAIL');
 
         parent::configure();
+    }
+
+    #[Override]
+    protected function getDefaultRows(): array
+    {
+        return [
+            $this->getHelpText(),
+            $this->getEmailField(),
+        ];
     }
 
     protected function getHelpText(): ?Stringable